test(vetting): Wave 5 — morphology mega-split (6 files) + Octave/MATLAB research
Split test_shape_morphology_2d.h (11 fns, 3 maps) into test_morphology_{common,
regression,skimage,matlab,fraclac}.h. Functions were kind-pure at function level
(only convex_hull interleaves 1 CIRCULARITY assert -> kept whole in skimage).
Byte-for-byte preserved; common header avoids ODR dup. Verified in conda build:
full gtest suite 696/696 -- NO TESTS DROPPED. Registry repointed per feature.

MIGRATION.md 5.13: Octave + image/statistics is a near-drop-in for the MATLAB
oracle (only graycoprops needs a ~15-line shim) -> matlab-vetted rows become
license-free; recommend an 'octave' token. (external-verif-nyxus investigation.)
