+++ tests/test_all.cc
@@ -10,7 +10,10 @@
 #include "test_hu_mechanics.h"
 #include "test_hounsfield_loader.h"
 #include "test_morphology_features.h"
-#include "test_shape_morphology_2d.h"
+#include "test_morphology_regression.h"
+#include "test_morphology_skimage.h"
+#include "test_morphology_matlab.h"
+#include "test_morphology_fraclac.h"
 #include "test_moments_skimage.h"
 #include "test_moments_regression.h"
 #include "test_2d_remaining_features.h"
